project practice/requirements.txt
# ── V1: YOLOv8 세그멘테이션 ───────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
ultralytics==8.4.14
opencv-python-headless==4.10.0.84
# numpy는 이미 설치됨 (2.4.2) — 별도 설치 불필요

# ── V2: Medical RAG (LangChain + ChromaDB) ────────────────────
# Python 3.13 호환 버전 (numpy<2.0 제한 없음)
langchain==0.3.27
langchain-core==0.3.83
langchain-community==0.3.27
langchain-chroma==0.2.6
langchain-huggingface==0.3.1
langchain-openai==0.3.35
langchain-text-splitters==0.3.11

# 임베딩 모델 (BGE-M3 로컬 실행)
sentence-transformers>=3.0.0

# PDF 파싱
pymupdf>=1.24.0

# OpenAI API
openai>=1.40.0

# 환경변수 관리
python-dotenv>=1.0.0

# ChromaDB
chromadb>=0.5.0
